修复电源按钮

+++ config.sh
@@ -12,11 +12,13 @@ cp -r config/.Xresources ~/
 cp -r config/autostart ~/.config/
 
 # 安装top bar的字体
-mkdir -p ~/.local/share/fonts
-wget https://github.com/ryanoasis/nerd-fonts/releases/download/v3.0.0/JetBrainsMono.zip
-unzip JetBrainsMono.zip -d $HOME/.local/share/fonts/JetBrainsMono
-fc-cache
-rm -rf JetBrain*
+if [ ! -d ~/.local/share/fonts/JetBrainsMono ]; then
+	mkdir -p ~/.local/share/fonts
+	wget https://github.com/ryanoasis/nerd-fonts/releases/download/v3.0.0/JetBrainsMono.zip
+	unzip JetBrainsMono.zip -d $HOME/.local/share/fonts/JetBrainsMono
+	fc-cache
+	rm -rf JetBrain*
+fi
 
 # top bar的更新查看器要用
 sudo apt install -y apt-show-versions
